Update battor_agent.exe binary in cloud storage to fix stability issues.
We have a very old version of the Windows battor_agent in cloud storage
that is quite unstable. This new version contains the new init retry patch
https://codereview.chromium.org/1991403002/ that further improves stability.
The init retry patch has not landed yet, but we want to deploy this binary
as soon as possible to make sure everything is stable.

BUG=catapult:#2381

Review-Url: https://codereview.chromium.org/2065863002