Update sheriff entity *before* adding test updates to taskqueue.
Currently, sheriffing rotations are stored like this:
 * There is a Sheriff entity with a list of test patterns it monitors
 * When a test is put(), we check it against the sheriff patterns, and
   update its sheriff property if it matches one.

This code forces the put() to happen on tests affected by a sheriffing
pattern. It looks like only some of the tests are updated, and I think
the reason is that some put()s in the queue get executed before the sheriff
property is saved. This re-orders the code to put() the sheriff first,
and to add a delay before executing task queue tasks.

BUG=chromium:621499

Review-Url: https://codereview.chromium.org/2085523002