[Telemetry] Add Benchmark.ShouldTearDownStateAfterEachStorySetRun
Defaults to True in order to reset the state and make individual story set
repeats more independent of each other. The intended effect is to average
out noise in measurements between repeats.

Long running benchmarks willing to stess test the browser and have it run
for long periods of time may switch this value to False.

BUG=catapult:#2294

Review-Url: https://codereview.chromium.org/2091103003