[telemetry] Add support for setting browser logging verbosity
This patch replaces the existing enable_logging boolean field in
BrowserOptions (--enable-browser-logging flag) to a logging_verbosity
integer field (--browser-logging-verbosity flag).

  BEFORE                   AFTER
  [no flag]                --browser-logging-verbosity none [or no flag]
  [NOT SUPPORTED]          --browser-logging-verbosity non-verbose
  --enable-browser-logging --browser-logging-verbosity verbose

Note that the old flag as well as the field are still supported, but
they're marked as deprecated.

Rationale: We want to be able to enable browser logging with custom
verbosity level (--v=0) rather than the default one (--v=1).

BUG=chromium:623058

Review-Url: https://codereview.chromium.org/2093923002