Implement composition-span as a BarChart
A Composition diagnostic semantically describes its containing Value as a sum of
Values. Composition-span portrays this relationship as a single column in a
stacked BarChart.

When a Composition diagnostic contains a Numeric (as opposed to a ScalarNumeric),
its sum statistic is used for the data point in the BarChart. I'm not sure whether
that should be parameterized or generalized somehow.

This necessitated fixing BarChart's handling of a single datum.
BarChart is also restructured as a RelatedValueMap, which also gains a helper method
add(value) => set(value.name, value).

Fixing Chart legends (moving them out of the chart area, allowing them to be
analysis-links, and adding checkboxes to optionally hide data series)
is left for another CL: https://codereview.chromium.org/2124113002

Review-Url: https://codereview.chromium.org/2120323002