Implement numeric-stats-span as a table of stats
Currently, histogram-span displays average and count (and they look a bit
rough); scalar-span displays average.
None of the other stats are displayed anywhere that I can see.

numeric-stats-span displays all of the stats computed by a Numeric in a table.

This CL makes histogram-span display the stats to the right of the BarChart.

As polishing side-effects, the "count" unit is added to represent whole numbers,
and the "nans" Numeric stat contains the number of NaNs added to the Numeric.

Screenshot (google-only): https://screenshot.googleplex.com/FDwAzvK2mYE.png

BUG=catapult:#2180

Review-Url: https://codereview.chromium.org/2100413002