Fix single BattOr case in BattOr-device mapping script.
What is supposed to happen is that if there's only one BattOr, we skip doing the whole mapping thing and instead just return the one BattOr. That is implemetned elsewhere in this script but is broken because it never gets to that point since GenerateSerialMap throws an error in that case. This CL fixes that error.

Review-Url: https://codereview.chromium.org/2128303002