Refactor value-set-table to facilitate merging/summarization.
Currently, value-set-table is constructed in a way that would make it a bit
difficult to add levels of subRows of merged Value groups.
This CL refactors value-set-table to make that a bit easier, and also refactors
the grouping methods in iteration_helpers, EventSet, and ValueSet.

No functionality changes, except that value-set-table is insignificantly less
efficient now that it iterates over values one extra time.

Review-Url: https://codereview.chromium.org/2137603003