Fix ValueSet.addValuesFromDicts
Currently, addValuesFromDicts() calls addValue(), which adds summary values for
Numerics to the ValueSet and to the Numeric's diagnostics...  but those summary
values should already exist in the valueDicts, in the Numeric's diagnostics at
least, even if the summary values were stripped out for some reason.

This CL makes addValuesFromDicts bypass adding the summary ScalarNumeric values.

Review-Url: https://codereview.chromium.org/2153673003