Always call setDataList when adding new submenus to test-picker.
The API for autocomplete-box is very strange. It has an attribute called
`datalist` and a property called `dataList`. When creating an
autocomplete-box it is assumed that the attribute will be set and then
setDataList will be called to set the property as well. It was not being
set here. I have no idea how this ever worked in 0.5, but this is the
correct use of the API. I call Polymer.dom.flush() before calling
setDataList because I am accessing the dom element for the
autocomplete-box, and it is created by a dom-repeat template that may
not update immediately.

This code needs to be killed with fire, but I prefer not to make major
changes during the polymer 1.0 update.

BUG=catapult:#2285
NOTRY=true
NOPRESUBMIT=true

Review-Url: https://codereview.chromium.org/2165453002