Add tracing/bin/strip_memory_infra_trace
We are running frequently in cases where traces w/ memory-infra
are too big and fail to load, as their uncompressed size > 256M.
While this is going to be fixed by bit.ly/TracingV2 we need a
short term solution to open those traces, to deal with cases like
crbug.com/623945.
This script strips a trace with memory-infra keeping only the last
two global dumps. It still guarantees that all processes for a dump
are seen, and tries to maintain the detailed dumps, if any.

BUG=chromium:630965

Review-Url: https://codereview.chromium.org/2184783002