Links the netlog viewer with catapult in order to set up tests. This does not create any working tests yet. This is simply the foundation needed to make use of Catapult's test infrastructure and align with its policies on directory structure.
A few important points on some of the changes in this CL:

1. A bin folder was created in order to hold any executable files. This
folder currently contains one file (run_dev_server_tests) which runs
any tests for the netlog viewer.

2. A build folder was created which contains a dev server config
script (netlog_viewer_dev_server_config). This is needed to get
the paths to all the test files.

3. All the source code was put into its own separate folder
(netlog_viewer/netlog_viewer) in order to make the directory more
similar to the rest of catapult and also improve readability of
the directory. The only files at the top level of the directory
are the README, OWNERS, and a netlog_viewer_project file.

4. The netlog_viewer_project file is useful for giving other files
knowledge of where certain files can be located in the directory
structure.

BUG=chromium:472699
NOPRESUBMIT=true

Review-Url: https://codereview.chromium.org/2187053002