[Tracing] Use GTest progress reporter in metric_runner
This CL updates metric_runner to use the GTest progress reporter, so that when
multiple traces are processed (which can take a while), the progress can be
observed.

Also, we were blowing up before if no values were produced by analyzing a given
trace, which is fixed in this CL (we will now see GTest's [ FAILED ] for failed
runs, so there is no need to throw an exception).

Review-Url: https://codereview.chromium.org/2208763002