Moved can_bisect from its own endpoint to graph_json.
Now that the bisect button has been removed from the group_report page,
the chart tooltip is the only consumer of /can_bisect. Instead of doing
an XHR to get the bisectability of a chart, just include the info in
the XHR for the chart data. This CL removes the /can_bisect endpoint
as it is no longer used.

BUG=catapult:#2552

Review-Url: https://codereview.chromium.org/2244783003