Make Slices subclassable (and overhaul subclassing/custom-subviews)
This modifies the custom type and subview creation logic so that slice's can
be subtyped in the same way we support for other events like AsyncSlices. And,
it modifies the the analysis view to allow custom analysis subviews, which allows
custom views for slices to be registered as well.

The first big part of this patch moves all preexisting subtype registries to be
created and managed by EventRegistry. This makes it a little more obvious when
a given subtype is doing the wrong thing with regard to custom registration, because
they had to write code to do so. The "copy and paste" event subclasses have to write
zero code to do subclassing. This isn't the ideal resting point, but it does get us
to a place where Slices can be extended.

The second part of this rewrites the AnalysisSubView and how it generates tabs.
Instead of the EventRegistry knowing about element names, views are now registered
by the views themselves, and the AnalysisSubView handles the rest. The ASV tracks
views by the Event they handle: if you have a custom event you want shown specially,
then you tell the ASV about that handle, and the tagname, and it will make sure to
show it. ASV works by tracking the inheritance structure of the Model objects: it
remembers which model objects are subclasses of which, and so when you give it a
selection of N arbitrary objects, it picks out the most specific set of views that
will show that selection.

ASV has one core restriction: it will never "split" the display of fundamental
subtypes between different views. Specifically, if you give ASV a thread slice, and
a V8ThreadSlice, and you have a custom subview for the V8 threadslice alongside the
generic subview for thread slices, ASV will only choose the ASV subview. This is
intentional: we do not want to confuse the user by having a given event show up
in multiple tabs, or similarly risk the ambiguity of having an event show up in only
one tab but then be missing in another.

BUG=catapult:#2672

Review-Url: https://codereview.chromium.org/2242733002