Update RAIL power metric to be consistent with other power metrics.
Makes the RAIL power metric generate a histogram with separate samples
for each instance of the RAIL stage in cases where there are multiple
RAIL stages with the same name. Also changes the name of the metric
output to match the format of other metrics (e.g. energy:<stage_name>)

Review-Url: https://codereview.chromium.org/2303003002